requirements.txt
Flask==2.3.3
Pillow==10.0.1
Werkzeug==2.3.7
pandas